refactor: First step to decouple the react app from feedcast client
